requirements.txt
fastapi
PyPDF2
langchain
google-generativeai
uvicorn
python-dotenv
